move functions related to user data to new module
